modularizing home component
