+++ electron-react-app-main/app/components/home/Home.tsx
@@ -5,13 +5,14 @@ import { Button } from '../ui/button'
 import Loading from '../loading'
 import { cn } from '@/lib/utils'
 import './styles.css'
+import NoIndex from '../NoIndex'
+import Results from '../Results'
 
 export default function Home() {
   const [query, setQuery] = useState("")
   const search = useConveyor("search")
   const [results, setResults] = useState<string[]>([])
   const [isIndexed, setIsIndexed] = useState<boolean | null>(null)
-  const [isIndexing, setIsIndexing] = useState<boolean>(false)
   
   const handleSearch = async () => {
     const res = await search.search(query)
@@ -22,54 +23,9 @@ export default function Home() {
     setIsIndexed(indexed)
   }
 
-  const handleStartIndexing = async () => {
-    const res = await search.openFileDialog()
-    
-    if (!res || res.length === 0) return //check to prevent indexing null 
-    
-    setIsIndexing(true)
-    try {
-      const indexRes = await search.index(res)
-      if (indexRes) {
-        setIsIndexed(true)
-      }
-    } catch (error) {
-      console.error('Error indexing files:', error)
-    } finally {
-      setIsIndexing(false)
-    }
-  }
-
   // Show loading while checking index
   if (isIndexed === null) {
     return <Loading onIndexComplete={handleIndexComplete} />
-  }
-
-  // Show scan prompt if not indexed
-  if (isIndexed) {
-    return (
-      <div className="welcome-content flex flex-col gap-5 items-center justify-center min-h-[400px]">
-        <div className="loading-container">
-          {isIndexing ? (
-            <>
-              <div className="spinner"></div>
-              <h2 className="loading-title">Indexing Files...</h2>
-              <p className="loading-text">
-                Please wait while we scan and index your files. This may take a few moments.
-              </p>
-            </>
-          ) : (
-            <>
-              <h2>No Index Found</h2>
-              <p>Scan your folders to enable search functionality.</p>
-              <Button onClick={handleStartIndexing} className="mt-4">
-                Start Scanning
-              </Button>
-            </>
-          )}
-        </div>
-      </div>
-    )
   }
 
   // Show searchbar if indexed
@@ -95,18 +51,12 @@ export default function Home() {
         "border-2 border-zinc-700/80 bg-zinc-800/60",
         "px-4 shadow-[0_0_0_1px_rgba(255,255,255,0.03)]",
       )}>
-        Results here
+        {!isIndexed ? (
+          <NoIndex />
+        ) : (
+            <Results results={results}  query={query} />
+        )}
       </div>
-      
-      {/*{results.length > 0 && (
-        <div className="results-container">
-          {results.map((result, idx) => (
-            <div key={idx} className="result-item">
-              {result}
-            </div>
-          ))}
-        </div>
-      )}*/}
     </div>
   )
 }
